refactor(request): split fetch-helpers along its natural seams
Splits lib/request/fetch-helpers.ts (1,497 lines) into four focused
modules, moving code verbatim:

- token-refresh.ts (147 lines): shouldRefreshToken,
  refreshAndUpdateToken, plus private CodexAuthSetter,
  isRetryableRefreshFailure, isRetryableAuthSetterError
- error-classification.ts (366 lines): isEntitlementError,
  isWorkspaceDisabledError, createEntitlementErrorResponse, the
  unsupported-Codex-model detection/fallback chain and its private
  helpers (canonicalizeModelName, normalizeFallbackChain, patterns)
- url-rewriting.ts (199 lines): extractRequestUrl, rewriteUrlForCodex,
  proxy resolution (resolveProxyUrlForRequest, applyProxyCompatibleInit,
  closeSharedProxyDispatchers) and the shared dispatcher cache
- headers.ts (99 lines): createCodexHeaders (all overloads) and the
  RFC 8594 deprecation/sunset header logging

fetch-helpers.ts (now 756 lines) remains the re-export facade for every
previously-exported symbol, so no importer changes anywhere. Response
handling (handleErrorResponse/handleSuccessResponse), rate-limit
parsing, and transformRequestForCodex stay in fetch-helpers.ts as they
do not fit a seam cleanly. Runtime export surface verified identical
before/after; no import cycles introduced. Zero behavior change.

Audit roadmap §4.1.2.
